pyproject.toml
[tool.poetry]
name = "ffscrapper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
pyqt6 = "^6.4.2"
requests = "^2.28.2"
pandas = "^1.5.3"
bs4 = "^0.0.1"
xlsxwriter = "^3.0.9"
pillow = "^9.4.0"
xlrd = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
